Change data type in service code

+++ src/CreditPaymentService.java
@@ -1,5 +1,5 @@
 public class CreditPaymentService {
-    public double payment(double percent, double amount, double term) { ///  процент по кредиту, сумма, количество лет
+    public int payment(double percent, double amount, double term) { ///  процент по кредиту, сумма, количество лет
         double rPercent = percent / 100;
         double termToMonth = term * 12;
         double monthPercent = rPercent / 12;
